Move BackendController into own namespace (allows to use prefix)
