+++ core-bundle/src/Controller/Backend/BackendController.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
  * @license LGPL-3.0-or-later
  */
 
-namespace Contao\CoreBundle\Controller;
+namespace Contao\CoreBundle\Controller\Backend;
 
 use Contao\BackendAlerts;
 use Contao\BackendConfirm;
@@ -21,6 +21,7 @@ use Contao\BackendMain;
 use Contao\BackendPage;
 use Contao\BackendPassword;
 use Contao\BackendPopup;
+use Contao\CoreBundle\Controller\AbstractController;
 use Contao\CoreBundle\Picker\PickerBuilderInterface;
 use Contao\CoreBundle\Picker\PickerConfig;
 use Symfony\Component\HttpFoundation\InputBag;
@@ -31,8 +32,6 @@ use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 
 /**
- * @Route(defaults={"_scope" = "backend", "_token_check" = true})
- *
  * @internal
  */
 class BackendController extends AbstractController
@@ -77,7 +76,7 @@ class BackendController extends AbstractController
     /**
      * Symfony will un-authenticate the user automatically by calling this route.
      *
-     * @Route("/contao/logout", name="contao_backend_logout")
+     * @Route("/logout", name="contao_backend_logout")
      */
     public function logoutAction(): RedirectResponse
     {
@@ -85,7 +84,7 @@ class BackendController extends AbstractController
     }
 
     /**
-     * @Route("/contao/password", name="contao_backend_password")
+     * @Route("/password", name="contao_backend_password")
      */
     public function passwordAction(): Response
     {
@@ -97,7 +96,7 @@ class BackendController extends AbstractController
     }
 
     /**
-     * @Route("/contao/confirm", name="contao_backend_confirm")
+     * @Route("/confirm", name="contao_backend_confirm")
      */
     public function confirmAction(): Response
     {
@@ -109,7 +108,7 @@ class BackendController extends AbstractController
     }
 
     /**
-     * @Route("/contao/file", name="contao_backend_file")
+     * @Route("/file", name="contao_backend_file")
      */
     public function fileAction(): Response
     {
@@ -121,7 +120,7 @@ class BackendController extends AbstractController
     }
 
     /**
-     * @Route("/contao/help", name="contao_backend_help")
+     * @Route("/help", name="contao_backend_help")
      */
     public function helpAction(): Response
     {
@@ -133,7 +132,7 @@ class BackendController extends AbstractController
     }
 
     /**
-     * @Route("/contao/page", name="contao_backend_page")
+     * @Route("/page", name="contao_backend_page")
      */
     public function pageAction(): Response
     {
@@ -145,7 +144,7 @@ class BackendController extends AbstractController
     }
 
     /**
-     * @Route("/contao/popup", name="contao_backend_popup")
+     * @Route("/popup", name="contao_backend_popup")
      */
     public function popupAction(): Response
     {
@@ -157,7 +156,7 @@ class BackendController extends AbstractController
     }
 
     /**
-     * @Route("/contao/alerts", name="contao_backend_alerts")
+     * @Route("/alerts", name="contao_backend_alerts")
      */
     public function alertsAction(): Response
     {
@@ -175,7 +174,7 @@ class BackendController extends AbstractController
      *
      * @throws BadRequestHttpException
      *
-     * @Route("/contao/picker", name="contao_backend_picker")
+     * @Route("/picker", name="contao_backend_picker")
      */
     public function pickerAction(Request $request): RedirectResponse
     {
